requirements.txt
streamlit==1.38.0
langchain==0.2.16
langchain-community==0.2.16
langchain-groq==0.1.9
langchain-core==0.2.38
langchain-chroma==0.1.3
langchain-text-splitters==0.2.4
python-dotenv==1.0.1
langchain-huggingface==0.0.3
sentence-transformers==3.0.1
beautifulsoup4==4.12.3